% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assume_role.R
\name{assume_role_with_web_identity}
\alias{assume_role_with_web_identity}
\title{Assume Role with AWS Web Identity}
\usage{
assume_role_with_web_identity(
  role_arn,
  token_file,
  base_url = Sys.getenv("AWS_STS_ENDPOINT", "https://sts.amazonaws.com"),
  session_name = NULL,
  duration = 3600,
  version = "2011-06-15",
  verbose = getOption("verbose", FALSE)
)
}
\arguments{
\item{role_arn}{A character string containing the AWS Role Amazon Resource Name (ARN). This specifies the permissions you have to access other AWS services.}

\item{token_file}{A character string containing a path to a Web Identity Token file.}

\item{base_url}{The AWS STS endpoint to use to retrieve your credentials from.}

\item{session_name}{A character string optionally specifying the name.}

\item{duration}{The expiry time on the retrieved credentials.}

\item{version}{The AWS STS specification version to use.}

\item{verbose}{A logical indicating whether to be verbose.}
}
\description{
Assume a role from a provided Web Identity Token and ARN using AWS Secure Token Service (STS).
}
